Be explicit for AttachImageBuilderVolume.__str__
InjectableModel.__str__ looks up several attributes to try and find a
good description for the model.  AttachImageBuilderVolume is both an
InjectableModel and a BaseCustomization.
BaseCustomization.__getattr__ used to call __str__ in constructing the
error to raise when an attribute is not found.  That created a loop so
instead of getting AttributeError, we got a recursion depth exceeded.

So every time an AttachImageBuilderVolume needed to describe itself (in the setup_task logging), we got an error.

Fix both on the BaseCustomization side to use repr rather than str,
and here to give an explicit __str__.