Add AwsPrivateHostedZone class for creating private zones
Private hosted zones don't have DelegationSet.
  Recreate find_by_id which skip setting the nameservers attr since there are not nameservers for a private hosted zone.

Add a delete function so that we can delete a Hosted Zone.
Add a clear function so that when deleting a zone all
non-required records can be deleted before calling delete on the zone.

Add a check in the do_create for private hosted zones to ensure that enableDnsHostnames is enabled.

Add test with retry logic to ensure the txt record is resolvable from a
vm in the VPC associated to the private hosted zone.

Add dns_enable_hostnames to the AwsVirtualPrivateCloud class.
The default VPC in aws is forced to be readonly to prevent accidental
changes to its attributes.

Fixes GH-19