Enhance the pre-commit hook to not allow unlinted files in on accident.
Previously we had a situation that could occur which would allow for unlinted files to be commited if they previously failed the pre-commit hook then were corrected but not staged. This would effectivly allow an unlinted file to be commited.

This change will create a temp dir, and copy the files to the temp dir this way we only lint the files that have been staged. If it fails the temp dir is left intact so that it can be inspected by the user. Once it gets a successful run it'll clean itself up.